Converted categoriesContext to use redux and removed all traces of categories context from project
